add JWT

+++ src/users/users.controller.ts
@@ -6,41 +6,63 @@ import {
   Param,
   Put,
   Delete,
+  UseGuards,
+  ForbiddenException,
+  Request,
 } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { CreateUserDto } from './dtos/createUser.dto';
 import { UpdateUserDto } from './dtos/updateUser.dto';
 import { LoginDto } from './dtos/login.dto';
+import { AuthService } from 'src/auth/auth.service';
+import { JwtAuthGuard } from 'src/auth/auth.guard';
 
 @Controller('api/users')
 export class UsersController {
-  constructor(private readonly usersService: UsersService) {}
+  constructor(
+    private readonly usersService: UsersService,
+    private readonly authService: AuthService,
+  ) {}
 
   @Post('signup')
   async createUser(@Body() body: CreateUserDto) {
     return this.usersService.createUser(body);
   }
 
-  // TODO: JWT later
+  @UseGuards(JwtAuthGuard)
   @Get(':id')
-  async getUser(@Param('id') id: string) {
+  async getUser(@Param('id') id: string, @Request() req) {
+    if (req.user.userId !== id) {
+      throw new ForbiddenException('Access denied');
+    }
     return this.usersService.getUserById(id);
   }
 
-  // TODO: JWT later
+  @UseGuards(JwtAuthGuard)
   @Put(':id')
-  async updateUser(@Param('id') id: string, @Body() body: UpdateUserDto) {
+  async updateUser(
+    @Param('id') id: string,
+    @Body() body: UpdateUserDto,
+    @Request() req,
+  ) {
+    if (req.user.userId !== id) {
+      throw new ForbiddenException('Access denied');
+    }
     return this.usersService.updateUser(id, body);
   }
 
-  // TODO: JWT later
+  @UseGuards(JwtAuthGuard)
   @Delete(':id')
-  async removeUser(@Param('id') id: string) {
+  async removeUser(@Param('id') id: string, @Request() req) {
+    if (req.user.userId !== id) {
+      throw new ForbiddenException('Access denied');
+    }
     return this.usersService.removeUser(id);
   }
 
   @Post('login')
   async login(@Body() body: LoginDto) {
-    return this.usersService.login(body);
+    const user = await this.usersService.login(body);
+    return this.authService.getJwt(user);
   }
 }
